computer_institute/requirements.txt
gradio
mysql-connector-python
python-dotenv
